requested changes

+++ src/main/java/org/prebid/server/proto/openrtb/ext/request/connatix/ExtImpConnatix.java
@@ -3,6 +3,7 @@ package org.prebid.server.proto.openrtb.ext.request.connatix;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Value;
+import java.math.BigDecimal;
 
 @Builder(toBuilder = true)
 @Value(staticConstructor = "of")
@@ -12,6 +13,6 @@ public class ExtImpConnatix {
     String placementId;
 
     @JsonProperty("viewabilityPercentage")
-    Float viewabilityPercentage;
+    BigDecimal viewabilityPercentage;
 
 }
